+++ tasmota/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add support for Jarolift rollers by Keeloq algorithm
 - Add Zigbee features and improvements and remove support for Zigbee commands starting with ``Zigbee...``
 - Add support for MaxBotix HRXL-MaxSonar ultrasonic range finders by Jon Little (#7814)
+- Add support for Romanian language translations by Augustin Marti
 
 ### 8.1.0.9 20200220